@@ -4,41 +4,44 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "next",
-    "build": "next build && next export",
-    "type-check": "tsc"
+    "dev": "vite",
+    "prebuild": "tsc",
+    "build": "vite build && vite build --ssr src/entry-ssg.tsx --outDir generated/ssg && NODE_ENV=production node scripts/ssg"
   },
   "keywords": [],
   "author": "iMasanari",
   "license": "MIT",
   "dependencies": {
     "@emotion/cache": "^11.7.1",
-    "@emotion/react": "^11.7.1",
+    "@emotion/react": "^11.8.2",
     "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.6.0",
+    "@emotion/styled": "^11.8.1",
     "@mdx-js/react": "^1.6.22",
-    "@mui/icons-material": "^5.3.0",
-    "@mui/material": "^5.3.0",
-    "@xmldom/xmldom": "^0.8.0",
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.5.1",
+    "@xmldom/xmldom": "^0.8.1",
     "action-reducer": "^0.3.1",
-    "jspdf": "^2.5.0",
+    "jspdf": "^2.5.1",
     "keycode-js": "^3.1.0",
-    "next": "^12.0.8",
     "promise-worker": "^2.0.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-image-crop": "^9.0.5",
+    "react-image-crop": "^9.1.1",
     "svg2pdf.js": "^2.2.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.7",
+    "@babel/core": "^7.17.7",
     "@babel/preset-react": "^7.16.7",
-    "@emotion/babel-preset-css-prop": "^11.2.0",
-    "@mdx-js/loader": "^1.6.22",
-    "@next/mdx": "^12.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@emotion/babel-plugin": "^11.7.2",
+    "@types/node": "^17.0.21",
+    "@types/react": "^17.0.40",
+    "@types/react-dom": "^17.0.13",
     "@types/react-image-crop": "^8.1.3",
-    "typescript": "^4.5.4"
+    "@vitejs/plugin-react": "^1.2.0",
+    "typescript": "^4.6.2",
+    "vite": "^2.8.6",
+    "vite-plugin-ejs": "^1.4.4",
+    "vite-plugin-mdx": "^3.5.10",
+    "vite-tsconfig-paths": "^3.4.1"
   }
 }
