@@ -1,8 +1,9 @@
 import PromiseWorker from 'promise-worker'
 import { Option } from './worker'
+import PdfWorker from './worker?worker'
 
-const ps = typeof window === 'object' ?
-  new PromiseWorker(new Worker(new URL('./worker.ts', import.meta.url)))
+const ps = typeof window === 'object'
+  ? new PromiseWorker(new PdfWorker())
   : null!
 
 export const createPdfFile = (option: Option) =>
