@@ -1,9 +1,6 @@
 import { css, Global, Theme } from '@emotion/react'
 import { CssBaseline } from '@mui/material'
-import Head from 'next/head'
 import React, { useReducer } from 'react'
-import favicon from '../assets/favicon.png'
-import icon from '../assets/icon.png'
 import Header from '~/common/layouts/Header'
 import Cards from '~/features/cards/Cards'
 import cardsReducer, { CardsState } from '~/features/cards/cardsReducer'
@@ -23,7 +20,7 @@ const globalStyle = css`
   }
   html,
   body,
-  #__next {
+  #app {
     @media (min-width: 600px) {
       height: 100%;
     }
@@ -82,18 +79,6 @@ const Index = () => {
 
   return (
     <div css={appStyle}>
-      <Head>
-        <title>プロキシカード印刷</title>
-        <meta name="description" content="カードゲームのプロキシ(コピーカード)を簡単に印刷できるWebアプリ。PC・スマホの両方に対応しています。" />
-        <meta property="og:title" content="プロキシカード印刷" />
-        <meta property="og:type" content="website" />
-        <meta property="og:description" content="カードゲームのプロキシ(コピーカード)を簡単に印刷できるWebアプリです。PC・スマホの両方に対応しています。" />
-        <meta property="og:url" content="https://imasanari.github.io/proxy-card-print/" />
-        <meta property="og:image" content={`https://imasanari.github.io/proxy-card-print${icon.src}`} />
-        <meta name="twitter:card" content="summary" />
-        <link rel="icon" href={favicon.src} />
-        <link rel="apple-touch-icon" href={icon.src} />
-      </Head>
       <CssBaseline />
       <Global styles={globalStyle} />
       <Header />
